+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `Super::Assets.use_sprockets`. It does what it says it does now. It used to do
   nothing.
+* Handling of prerelease versions of Sprockets and Webpacker.
 
 ## [0.0.13] - 2021-04-16